+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`2022011100...main`](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/compare/2022011100...main).
 
+### Changed
+
+- Started displaying the Matrix chat topic below the name of the chat in course view ([#62](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/merge_requests/62))
+
 ## [`2022011101`](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/tags/2022011101)
 
 For a full diff see [`2022011100...2022011101`](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/compare/2022011100...2022011101).